:sparkles: todo is now done

+++ src/todo/todo.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, FindManyOptions } from 'typeorm';
+import { Repository, FindManyOptions, FindOneOptions } from 'typeorm';
 
 import { ITodo } from '../common/types';
 import { Todo } from './todo.entity';
@@ -13,16 +13,26 @@ export class TodoService {
     private userService: UserService,
   ) {}
 
-  findAll(options?: FindManyOptions): Promise<Todo[]> {
-    return this.todoRepository.find(options);
+  async findAll(options?: FindManyOptions): Promise<Todo[]> {
+    return await this.todoRepository.find(options);
   }
 
-  findOne(id?: string, options?: any): Promise<Todo> {
-    return this.todoRepository.findOne(id, options);
+  async findOne(id?: number, options?: FindOneOptions): Promise<Todo> {
+    const todo: Todo = await this.todoRepository.findOne(id, options);
+
+    if (!todo) {
+      throw new Error('Todo not found');
+    }
+
+    return todo;
   }
 
   async create({title, text, userId}: ITodo): Promise<Todo> {
     const user = await this.userService.findOne(userId);
+
+    if (!user) {
+      throw new Error('User not found');
+    }
 
     const todo: Todo = await this.todoRepository.create({
       title,
@@ -35,7 +45,38 @@ export class TodoService {
     return todo;
   }
 
+  async update({ id, title, text }: ITodo): Promise<Todo> {
+    const todo: Todo = await this.todoRepository.findOne(id);
+
+    if (!todo) {
+      throw new Error('Todo not found');
+    }
+
+    await this.todoRepository.update(id, {
+      text,
+      title,
+    });
+
+    return await this.todoRepository.findOne({
+      where: { id },
+      relations: ['items', 'user'],
+    });
+  }
+
   async remove(id: number): Promise<void> {
+    const todo: Todo = await this.todoRepository.findOne({
+      where: { id },
+      relations: ['items', 'user'],
+    });
+
+    if (!todo) {
+      throw new Error('Todo not found');
+    }
+
+    if (todo.items && todo.items.length > 0) {
+      throw new Error('Cannot delete todo with items');
+    }
+
     await this.todoRepository.delete(id);
   }
 }
